requirements.txt
fastapi
uvicorn
requests
dvc
GitPython
torch
torchvision
mlxtend